% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{analize_blocks}
\alias{analize_blocks}
\title{Analysis of honeycomb selection design based on blocks of unique nearby entries.}
\usage{
analize_blocks(
  Main_Data_Frame = NULL,
  observation = NULL,
  row_element = NULL,
  plant_element = NULL,
  CRS,
  rep_unrep
)
}
\arguments{
\item{observation}{A vector containing the observations.}

\item{row_element}{The row of the element which the block it belongs to will be displayed.}

\item{plant_element}{The position of the element in the row which the block it belongs to will be displayed.}

\item{CRS}{Number of top plants used for the CRS index.}

\item{rep_unrep}{Replicated of unreplicated design.}
}
\value{
A dataframe.
}
\description{
A Function to analyze blocks of entries. The "central" plant in each position is not calculated.
}
\keyword{internal}
