% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{HSD01}
\alias{HSD01}
\title{Construction of the honeycomb selection design with one control.}
\usage{
HSD01(K, rows, plpr, distance, poly = TRUE)
}
\arguments{
\item{K}{The K parameter.}

\item{rows}{The number of rows.}

\item{plpr}{The number of plants per row.}

\item{distance}{Distance between plants in meters.}

\item{poly}{If TRUE  the polygon pattern is displayed.}
}
\value{
A dataframe.
}
\description{
This function creates a data frame of an  honeycomb selection design  (one entry, one control).
}
\examples{
HSD01(1,10,10,1) 
}
\references{
Fasoula V. (2013). Prognostic Breeding: A New Paradigm for Crop Improvement. Plant Breeding Reviews 37: 297-347. 10.1002/9781118497869.ch6. \doi{10.1002/9781118497869.ch6}

Fasoula V.A., and Tokatlidis I.S. (2012). Development of crop cultivars by honeycomb breeding. Agronomy for Sustainable Development 32:161–180. 10.1007/s13593-011-0034-0 \doi{10.1007/s13593-011-0034-0}

Fasoulas A.C., and Fasoula V.A. (1995). Honeycomb selection designs. In J. Janick (ed.). Plant Breeding Reviews 13: 87-139.  \doi{10.1002/9780470650059.ch3}

Tokatlidis I. (2016). Sampling the spatial heterogeneity of the honeycomb model in maize and wheat breeding trials: Analysis of secondary data compared to popular classical designs. Experimental Agriculture, 52(3), 371-390. \doi{10.1017/S0014479715000150}

Tokatlidis I., and Vlachostergios D. (2016). Sustainable Stewardship of the Landrace Diversity. Diversity 8(4):29. \doi{10.3390/d8040029}
}
