% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_convert}
\alias{plot_convert}
\title{This function returns a plot.}
\usage{
plot_convert(dataf, poly = TRUE, y_rev = TRUE, x_rev = FALSE, rep_unrep = NULL)
}
\arguments{
\item{dataf}{Data frame containing information about the experiment.}

\item{poly}{If TRUE set the polygon pattern.}

\item{y_rev}{Reverse the y axis.}

\item{x_rev}{Reverse the x axis.}

\item{rep_unrep}{Replicated or unreplicated selection design.}
}
\value{
A dataframe.
}
\description{
It prints a graphic.
}
\keyword{internal}
