% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{return_ungrouped}
\alias{return_ungrouped}
\title{This function returns only the ungrouped replicated selection designs.}
\usage{
return_ungrouped(R_gen)
}
\arguments{
\item{R_gen}{A single number or vector containing the replicated selection designs for testing.}
}
\value{
A dataframe.
}
\description{
It calls the check for R function and keeps only the Ungrouped selection designs.
}
\keyword{internal}
