% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{test_for_R}
\alias{test_for_R}
\title{Tests if a selection design exists and returns its X and Y values.}
\usage{
test_for_R(R_gen)
}
\arguments{
\item{R_gen}{A single number or vector containing the replicated selection designs for testing.}
}
\value{
A dataframe.
}
\description{
It is used to return the X and Y values of a replicated selection design if it exists.
}
\keyword{internal}
