% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheat_data.R
\docType{data}
\name{wheat_data}
\alias{wheat_data}
\title{A dataset}
\format{
\describe{
\item{wheat_data$main_spike_weight}{The weight (g) of the main spike of a single plant.}
\item{wheat_data$tillers_spike_weight}{The weight (g) of tillers' spikes of a single plant.}
\item{wheat_data$total_yield}{The total yield (g) of a single plant.}

}
}
\usage{
wheat_data
}
\description{
A dataset containing observations from an R7 honeycomb selection design.
}
\keyword{datasets}
